Level2의 10번 요구사항 수정 - % 연산 추가

+++ src/main/java/calculator/AddOperator.java
@@ -1,7 +1,7 @@
 package calculator;
 
 public class AddOperator implements Operator {
-    public double operate(int num1, int num2) {
-        return (double)num1 + num2;
+    public int operate(int num1, int num2) {
+        return num1 + num2;
     }
 }
