+++ src/main/java/calculator/Operator.java
@@ -1,5 +1,5 @@
 package calculator;
 
 public interface Operator {
-    double operate(int num1, int num2);
+    int operate(int num1, int num2);
 }
